Refactor explicit.MDP, split generic part (independent of data type for transition probabilities) into MDPGeneric
This allows reuse of some methods in the exact / parametric engine.
